refactor(db): Split index_manager.rs into directory module
