+++ crates/db/src/index_manager/mod.rs
@@ -5,9 +5,11 @@
 /// - Dropping indexes (with entry cleanup)
 /// - Loading index instances from schema
 /// - Maintaining indexes during document mutations
+mod value_extraction;
+
 use crate::error::{Error, Result};
 use datastore::NamespaceView;
-use document::{Document, NormalValue};
+use document::Document;
 use schema::{CollectionVersion, FieldDescription, IndexDescription, IndexedFieldDescription};
 use std::collections::HashMap;
 use storage::corekv::Key;
@@ -90,7 +92,6 @@ impl IndexManager {
     pub fn from_collection(collection_short_id: u32, schema: &CollectionVersion) -> Result<Self> {
         let mut manager = Self::new(collection_short_id);
         for desc in &schema.indexes {
-            // Validate index configuration
             if desc.fields.is_empty() {
                 return Err(Error::Other(format!(
                     "index '{}' in schema has no fields",
@@ -138,7 +139,6 @@ impl IndexManager {
         unique: bool,
         schema_fields: &[FieldDescription],
     ) -> Result<IndexDescription> {
-        // Auto-generate name if empty (matches Go behavior)
         let name = if name.is_empty() {
             let first_field = fields.first().map(|f| f.name.as_str()).unwrap_or("unknown");
             let existing: Vec<String> = self.indexes.keys().cloned().collect();
@@ -147,7 +147,6 @@ impl IndexManager {
             name
         };
 
-        // Validate index name
         if !is_valid_index_name(&name) {
             return Err(Error::Other(format!(
                 "index with invalid name. Name: {}",
@@ -155,7 +154,6 @@ impl IndexManager {
             )));
         }
 
-        // Check if index already exists
         if self.indexes.contains_key(&name) {
             return Err(Error::Other(format!(
                 "index with name already exists. Name: {}",
@@ -163,14 +161,12 @@ impl IndexManager {
             )));
         }
 
-        // Validate fields
         if fields.is_empty() {
             return Err(Error::Other(
                 "index must have at least one field".to_string(),
             ));
         }
 
-        // Reject CRDT counter fields (matches Go's NewCollectionIndex validation)
         for field in &fields {
             if let Some(schema_field) = schema_fields.iter().find(|f| f.name == field.name) {
                 if schema_field.crdt_type.is_counter() {
@@ -182,10 +178,8 @@ impl IndexManager {
             }
         }
 
-        // Generate a new index ID
         let index_id = self.next_index_id(datastore).await?;
 
-        // Create the index description
         let desc = IndexDescription {
             name: name.clone(),
             id: index_id,
@@ -193,7 +187,6 @@ impl IndexManager {
             unique,
         };
 
-        // Create the index instance
         let index = IndexType::new(self.collection_short_id, desc.clone());
         self.indexes.insert(name, index);
 
@@ -218,8 +211,6 @@ impl IndexManager {
     pub async fn drop_index(&mut self, datastore: &NamespaceView, name: &str) -> Result<bool> {
         match self.indexes.remove(name) {
             Some(index) => {
-                // Create a mutable namespace view for deletion
-                // We need to use the underlying store operations
                 index
                     .remove_all(&mut datastore.clone())
                     .await
@@ -243,7 +234,6 @@ impl IndexManager {
         let seq_key = IndexIDSequenceKey::new(format!("{}", self.collection_short_id));
         let key_bytes = seq_key.bytes();
 
-        // Get current sequence value
         let current = match datastore.get(&key_bytes).await.map_err(Error::Storage)? {
             Some(bytes) => {
                 if bytes.len() == 4 {
@@ -255,7 +245,6 @@ impl IndexManager {
             None => 0,
         };
 
-        // Increment and store
         let next_id = current + 1;
         datastore
             .set(&key_bytes, &next_id.to_be_bytes())
@@ -297,16 +286,13 @@ impl IndexManager {
             let doc_id = match doc.id() {
                 Some(id) => id.to_string(),
                 None => {
-                    // Document without ID cannot be indexed - skip with warning
                     skipped_count += 1;
                     continue;
                 }
             };
 
-            // Extract field values for the index (may return multiple value sets for arrays)
             let value_sets = self.extract_index_values(doc, index.description(), schema)?;
 
-            // Save all value sets to index (one entry per array element combination)
             for values in &value_sets {
                 index
                     .save(&mut mutable_datastore, &doc_id, values)
@@ -339,7 +325,6 @@ impl IndexManager {
 
         for index in self.indexes.values() {
             let value_sets = self.extract_index_values(doc, index.description(), schema)?;
-            // Save all value sets (one entry per array element combination)
             for values in &value_sets {
                 index
                     .save(&mut mutable_datastore, &doc_id, values)
@@ -374,16 +359,13 @@ impl IndexManager {
             let old_value_sets = self.extract_index_values(old_doc, index.description(), schema)?;
             let new_value_sets = self.extract_index_values(new_doc, index.description(), schema)?;
 
-            // Only update if value sets changed
             if old_value_sets != new_value_sets {
-                // Delete all old entries
                 for old_values in &old_value_sets {
                     index
                         .delete(&mut mutable_datastore, &doc_id, old_values)
                         .await
                         .map_err(Error::Storage)?;
                 }
-                // Save all new entries
                 for new_values in &new_value_sets {
                     index
                         .save(&mut mutable_datastore, &doc_id, new_values)
@@ -412,7 +394,6 @@ impl IndexManager {
 
         for index in self.indexes.values() {
             let value_sets = self.extract_index_values(doc, index.description(), schema)?;
-            // Delete all value set entries (one per array element combination)
             for values in &value_sets {
                 index
                     .delete(&mut mutable_datastore, &doc_id, values)
@@ -422,247 +403,6 @@ impl IndexManager {
         }
 
         Ok(())
-    }
-
-    /// Extract field values from a document for indexing.
-    ///
-    /// # Multi-Value Indexing (Arrays)
-    ///
-    /// When a field contains an array, multiple index entries are created - one per
-    /// array element. For composite indexes with multiple array fields, the Cartesian
-    /// product of all array elements is generated.
-    ///
-    /// Example: For document `{tags: ["a", "b"], categories: ["x", "y"]}` with a
-    /// composite index on `(tags, categories)`, four index entries are created:
-    /// `("a", "x")`, `("a", "y")`, `("b", "x")`, `("b", "y")`
-    ///
-    /// # Null Handling
-    ///
-    /// If a document is missing a field that is part of an index, the value is
-    /// indexed as `NormalValue::Null`. This is intentional for nullable fields
-    /// and allows documents with missing optional fields to be indexed.
-    ///
-    /// For unique indexes, multiple documents with NULL values for the same
-    /// indexed field will all be indexed (NULL is not considered equal to NULL
-    /// for uniqueness purposes).
-    fn extract_index_values(
-        &self,
-        doc: &Document,
-        index_desc: &IndexDescription,
-        schema: &CollectionVersion,
-    ) -> Result<Vec<Vec<NormalValue>>> {
-        // Build a set of field names for O(1) lookup
-        let schema_fields: std::collections::HashSet<&str> =
-            schema.fields.iter().map(|f| f.name.as_str()).collect();
-
-        // Collect expanded values for each field (arrays become multiple values)
-        let mut field_value_sets: Vec<Vec<NormalValue>> =
-            Vec::with_capacity(index_desc.fields.len());
-
-        for field in &index_desc.fields {
-            // Validate that index field exists in schema (except for system fields)
-            if !field.name.starts_with('_') && !schema_fields.contains(field.name.as_str()) {
-                return Err(Error::Other(format!(
-                    "index '{}' references field '{}' which does not exist in schema",
-                    index_desc.name, field.name
-                )));
-            }
-
-            let value = doc.get(&field.name).cloned().unwrap_or(NormalValue::Null);
-
-            // Expand arrays into multiple values; scalars become single-element sets
-            let expanded = Self::expand_value_for_indexing(value);
-            field_value_sets.push(expanded);
-        }
-
-        // Compute Cartesian product of all field value sets
-        Ok(Self::cartesian_product(field_value_sets))
-    }
-
-    /// Expand a value for multi-value indexing.
-    ///
-    /// Arrays are expanded into their elements. Empty arrays result in a single
-    /// NULL value to ensure the document is still indexed.
-    fn expand_value_for_indexing(value: NormalValue) -> Vec<NormalValue> {
-        // Helper macro to handle array expansion with conversion to NormalValue
-        macro_rules! expand_array {
-            ($arr:expr, $variant:ident) => {
-                if $arr.is_empty() {
-                    vec![NormalValue::Null]
-                } else {
-                    $arr.iter()
-                        .map(|v| NormalValue::$variant(v.clone()))
-                        .collect()
-                }
-            };
-        }
-
-        // Helper macro for nillable arrays (Some/None)
-        macro_rules! expand_nillable_array {
-            ($opt:expr, $variant:ident) => {
-                match $opt {
-                    Some(arr) => {
-                        if arr.is_empty() {
-                            vec![NormalValue::Null]
-                        } else {
-                            arr.iter()
-                                .map(|v| NormalValue::$variant(v.clone()))
-                                .collect()
-                        }
-                    }
-                    None => vec![NormalValue::Null],
-                }
-            };
-        }
-
-        // Helper macro for arrays with nillable elements
-        macro_rules! expand_nillable_element_array {
-            ($arr:expr, $variant:ident) => {
-                if $arr.is_empty() {
-                    vec![NormalValue::Null]
-                } else {
-                    $arr.iter()
-                        .map(|v| match v {
-                            Some(val) => NormalValue::$variant(val.clone()),
-                            None => NormalValue::Null,
-                        })
-                        .collect()
-                }
-            };
-        }
-
-        match value {
-            // JSON values are expanded via leaf traversal
-            NormalValue::Json(_) => {
-                let leaves = value.json_leaves();
-                if leaves.is_empty() {
-                    // Empty JSON object/array produces single NULL entry
-                    vec![NormalValue::Null]
-                } else {
-                    leaves
-                }
-            }
-
-            // Non-array scalar types - single element
-            NormalValue::Null
-            | NormalValue::Bool(_)
-            | NormalValue::Int(_)
-            | NormalValue::Float64(_)
-            | NormalValue::Float32(_)
-            | NormalValue::String(_)
-            | NormalValue::Bytes(_)
-            | NormalValue::Time(_)
-            | NormalValue::Document(_)
-            | NormalValue::JsonLeaf(_)
-            | NormalValue::NillableBool(_)
-            | NormalValue::NillableInt(_)
-            | NormalValue::NillableFloat64(_)
-            | NormalValue::NillableFloat32(_)
-            | NormalValue::NillableString(_)
-            | NormalValue::NillableBytes(_)
-            | NormalValue::NillableTime(_)
-            | NormalValue::NillableDocument(_) => vec![value],
-
-            // Typed array types - expand to elements
-            NormalValue::BoolArray(ref arr) => expand_array!(arr, Bool),
-            NormalValue::IntArray(ref arr) => expand_array!(arr, Int),
-            NormalValue::Float64Array(ref arr) => expand_array!(arr, Float64),
-            NormalValue::Float32Array(ref arr) => expand_array!(arr, Float32),
-            NormalValue::StringArray(ref arr) => expand_array!(arr, String),
-            NormalValue::BytesArray(ref arr) => expand_array!(arr, Bytes),
-            NormalValue::TimeArray(ref arr) => expand_array!(arr, Time),
-            NormalValue::DocumentArray(ref arr) => {
-                if arr.is_empty() {
-                    vec![NormalValue::Null]
-                } else {
-                    arr.iter()
-                        .map(|v| NormalValue::Document(Box::new(v.clone())))
-                        .collect()
-                }
-            }
-            NormalValue::JsonArray(ref arr) => expand_array!(arr, Json),
-
-            // Nillable arrays (whole array can be null)
-            NormalValue::NillableBoolArray(ref opt) => expand_nillable_array!(opt, Bool),
-            NormalValue::NillableIntArray(ref opt) => expand_nillable_array!(opt, Int),
-            NormalValue::NillableFloat64Array(ref opt) => expand_nillable_array!(opt, Float64),
-            NormalValue::NillableFloat32Array(ref opt) => expand_nillable_array!(opt, Float32),
-            NormalValue::NillableStringArray(ref opt) => expand_nillable_array!(opt, String),
-            NormalValue::NillableBytesArray(ref opt) => expand_nillable_array!(opt, Bytes),
-            NormalValue::NillableTimeArray(ref opt) => expand_nillable_array!(opt, Time),
-            NormalValue::NillableDocumentArray(ref opt) => match opt {
-                Some(arr) => {
-                    if arr.is_empty() {
-                        vec![NormalValue::Null]
-                    } else {
-                        arr.iter()
-                            .map(|v| NormalValue::Document(Box::new(v.clone())))
-                            .collect()
-                    }
-                }
-                None => vec![NormalValue::Null],
-            },
-
-            // Arrays with nillable elements
-            NormalValue::NillableBoolElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Bool)
-            }
-            NormalValue::NillableIntElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Int)
-            }
-            NormalValue::NillableFloat64ElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Float64)
-            }
-            NormalValue::NillableFloat32ElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Float32)
-            }
-            NormalValue::NillableStringElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, String)
-            }
-            NormalValue::NillableBytesElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Bytes)
-            }
-            NormalValue::NillableTimeElementArray(ref arr) => {
-                expand_nillable_element_array!(arr, Time)
-            }
-            NormalValue::NillableDocumentElementArray(ref arr) => {
-                if arr.is_empty() {
-                    vec![NormalValue::Null]
-                } else {
-                    arr.iter()
-                        .map(|v| match v {
-                            Some(doc) => NormalValue::Document(Box::new(doc.clone())),
-                            None => NormalValue::Null,
-                        })
-                        .collect()
-                }
-            }
-        }
-    }
-
-    /// Compute Cartesian product of field value sets.
-    ///
-    /// Given `[[a, b], [x, y]]`, produces `[[a, x], [a, y], [b, x], [b, y]]`.
-    fn cartesian_product(sets: Vec<Vec<NormalValue>>) -> Vec<Vec<NormalValue>> {
-        if sets.is_empty() {
-            return vec![vec![]];
-        }
-
-        let mut result: Vec<Vec<NormalValue>> = vec![vec![]];
-
-        for set in sets {
-            let mut new_result = Vec::with_capacity(result.len() * set.len());
-            for combo in &result {
-                for val in &set {
-                    let mut new_combo = combo.clone();
-                    new_combo.push(val.clone());
-                    new_result.push(new_combo);
-                }
-            }
-            result = new_result;
-        }
-
-        result
     }
 
     /// Get all index names.
