+++ crates/db/src/index_manager/value_extraction.rs
@@ -1,0 +1,236 @@
+//! Index value extraction and Cartesian product logic.
+
+use crate::error::{Error, Result};
+use document::NormalValue;
+use schema::{CollectionVersion, IndexDescription};
+
+use super::IndexManager;
+
+impl IndexManager {
+    /// Extract field values from a document for indexing.
+    ///
+    /// # Multi-Value Indexing (Arrays)
+    ///
+    /// When a field contains an array, multiple index entries are created - one per
+    /// array element. For composite indexes with multiple array fields, the Cartesian
+    /// product of all array elements is generated.
+    ///
+    /// Example: For document `{tags: ["a", "b"], categories: ["x", "y"]}` with a
+    /// composite index on `(tags, categories)`, four index entries are created:
+    /// `("a", "x")`, `("a", "y")`, `("b", "x")`, `("b", "y")`
+    ///
+    /// # Null Handling
+    ///
+    /// If a document is missing a field that is part of an index, the value is
+    /// indexed as `NormalValue::Null`. This is intentional for nullable fields
+    /// and allows documents with missing optional fields to be indexed.
+    ///
+    /// For unique indexes, multiple documents with NULL values for the same
+    /// indexed field will all be indexed (NULL is not considered equal to NULL
+    /// for uniqueness purposes).
+    pub(crate) fn extract_index_values(
+        &self,
+        doc: &document::Document,
+        index_desc: &IndexDescription,
+        schema: &CollectionVersion,
+    ) -> Result<Vec<Vec<NormalValue>>> {
+        let schema_fields: std::collections::HashSet<&str> =
+            schema.fields.iter().map(|f| f.name.as_str()).collect();
+
+        let mut field_value_sets: Vec<Vec<NormalValue>> =
+            Vec::with_capacity(index_desc.fields.len());
+
+        for field in &index_desc.fields {
+            if !field.name.starts_with('_') && !schema_fields.contains(field.name.as_str()) {
+                return Err(Error::Other(format!(
+                    "index '{}' references field '{}' which does not exist in schema",
+                    index_desc.name, field.name
+                )));
+            }
+
+            let value = doc.get(&field.name).cloned().unwrap_or(NormalValue::Null);
+
+            let expanded = Self::expand_value_for_indexing(value);
+            field_value_sets.push(expanded);
+        }
+
+        Ok(Self::cartesian_product(field_value_sets))
+    }
+
+    /// Expand a value for multi-value indexing.
+    ///
+    /// Arrays are expanded into their elements. Empty arrays result in a single
+    /// NULL value to ensure the document is still indexed.
+    pub(super) fn expand_value_for_indexing(value: NormalValue) -> Vec<NormalValue> {
+        macro_rules! expand_array {
+            ($arr:expr, $variant:ident) => {
+                if $arr.is_empty() {
+                    vec![NormalValue::Null]
+                } else {
+                    $arr.iter()
+                        .map(|v| NormalValue::$variant(v.clone()))
+                        .collect()
+                }
+            };
+        }
+
+        macro_rules! expand_nillable_array {
+            ($opt:expr, $variant:ident) => {
+                match $opt {
+                    Some(arr) => {
+                        if arr.is_empty() {
+                            vec![NormalValue::Null]
+                        } else {
+                            arr.iter()
+                                .map(|v| NormalValue::$variant(v.clone()))
+                                .collect()
+                        }
+                    }
+                    None => vec![NormalValue::Null],
+                }
+            };
+        }
+
+        macro_rules! expand_nillable_element_array {
+            ($arr:expr, $variant:ident) => {
+                if $arr.is_empty() {
+                    vec![NormalValue::Null]
+                } else {
+                    $arr.iter()
+                        .map(|v| match v {
+                            Some(val) => NormalValue::$variant(val.clone()),
+                            None => NormalValue::Null,
+                        })
+                        .collect()
+                }
+            };
+        }
+
+        match value {
+            NormalValue::Json(_) => {
+                let leaves = value.json_leaves();
+                if leaves.is_empty() {
+                    vec![NormalValue::Null]
+                } else {
+                    leaves
+                }
+            }
+
+            NormalValue::Null
+            | NormalValue::Bool(_)
+            | NormalValue::Int(_)
+            | NormalValue::Float64(_)
+            | NormalValue::Float32(_)
+            | NormalValue::String(_)
+            | NormalValue::Bytes(_)
+            | NormalValue::Time(_)
+            | NormalValue::Document(_)
+            | NormalValue::JsonLeaf(_)
+            | NormalValue::NillableBool(_)
+            | NormalValue::NillableInt(_)
+            | NormalValue::NillableFloat64(_)
+            | NormalValue::NillableFloat32(_)
+            | NormalValue::NillableString(_)
+            | NormalValue::NillableBytes(_)
+            | NormalValue::NillableTime(_)
+            | NormalValue::NillableDocument(_) => vec![value],
+
+            NormalValue::BoolArray(ref arr) => expand_array!(arr, Bool),
+            NormalValue::IntArray(ref arr) => expand_array!(arr, Int),
+            NormalValue::Float64Array(ref arr) => expand_array!(arr, Float64),
+            NormalValue::Float32Array(ref arr) => expand_array!(arr, Float32),
+            NormalValue::StringArray(ref arr) => expand_array!(arr, String),
+            NormalValue::BytesArray(ref arr) => expand_array!(arr, Bytes),
+            NormalValue::TimeArray(ref arr) => expand_array!(arr, Time),
+            NormalValue::DocumentArray(ref arr) => {
+                if arr.is_empty() {
+                    vec![NormalValue::Null]
+                } else {
+                    arr.iter()
+                        .map(|v| NormalValue::Document(Box::new(v.clone())))
+                        .collect()
+                }
+            }
+            NormalValue::JsonArray(ref arr) => expand_array!(arr, Json),
+
+            NormalValue::NillableBoolArray(ref opt) => expand_nillable_array!(opt, Bool),
+            NormalValue::NillableIntArray(ref opt) => expand_nillable_array!(opt, Int),
+            NormalValue::NillableFloat64Array(ref opt) => expand_nillable_array!(opt, Float64),
+            NormalValue::NillableFloat32Array(ref opt) => expand_nillable_array!(opt, Float32),
+            NormalValue::NillableStringArray(ref opt) => expand_nillable_array!(opt, String),
+            NormalValue::NillableBytesArray(ref opt) => expand_nillable_array!(opt, Bytes),
+            NormalValue::NillableTimeArray(ref opt) => expand_nillable_array!(opt, Time),
+            NormalValue::NillableDocumentArray(ref opt) => match opt {
+                Some(arr) => {
+                    if arr.is_empty() {
+                        vec![NormalValue::Null]
+                    } else {
+                        arr.iter()
+                            .map(|v| NormalValue::Document(Box::new(v.clone())))
+                            .collect()
+                    }
+                }
+                None => vec![NormalValue::Null],
+            },
+
+            NormalValue::NillableBoolElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Bool)
+            }
+            NormalValue::NillableIntElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Int)
+            }
+            NormalValue::NillableFloat64ElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Float64)
+            }
+            NormalValue::NillableFloat32ElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Float32)
+            }
+            NormalValue::NillableStringElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, String)
+            }
+            NormalValue::NillableBytesElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Bytes)
+            }
+            NormalValue::NillableTimeElementArray(ref arr) => {
+                expand_nillable_element_array!(arr, Time)
+            }
+            NormalValue::NillableDocumentElementArray(ref arr) => {
+                if arr.is_empty() {
+                    vec![NormalValue::Null]
+                } else {
+                    arr.iter()
+                        .map(|v| match v {
+                            Some(doc) => NormalValue::Document(Box::new(doc.clone())),
+                            None => NormalValue::Null,
+                        })
+                        .collect()
+                }
+            }
+        }
+    }
+
+    /// Compute Cartesian product of field value sets.
+    ///
+    /// Given `[[a, b], [x, y]]`, produces `[[a, x], [a, y], [b, x], [b, y]]`.
+    pub(super) fn cartesian_product(sets: Vec<Vec<NormalValue>>) -> Vec<Vec<NormalValue>> {
+        if sets.is_empty() {
+            return vec![vec![]];
+        }
+
+        let mut result: Vec<Vec<NormalValue>> = vec![vec![]];
+
+        for set in sets {
+            let mut new_result = Vec::with_capacity(result.len() * set.len());
+            for combo in &result {
+                for val in &set {
+                    let mut new_combo = combo.clone();
+                    new_combo.push(val.clone());
+                    new_result.push(new_combo);
+                }
+            }
+            result = new_result;
+        }
+
+        result
+    }
+}
